refactor(viewlets): get ride of viewlets.xml to hide viewlets
I have change the way to hide viewlets to make it easier to
base a new theme on it. because hidding a viewlet for every
skin is not wanted, the goal is to hide viewlet to one skin.

To achieve this the common way is to use viewlets.xml skinname
attribute but this way you can't create a new theme based on
this one (so got a different skinanme) without copy/paste
the viewlets.xml.

Here I have take a new approch, where I create a new viewlet
with the same name but using a HiddenViewlet base class
which return an empty string; So now the layer is important
and you can create a new skin which inherits the layer.